Relocate some helper funcs in integration test
