+++ integration/main_test.go
@@ -75,3 +75,46 @@ func TestMain(m *testing.M) {
 
 	os.Exit(c)
 }
+
+// setup can be used to initialize things before integration tests start (as of now it only builds the services used by the integration tests so they can be referenced)
+func setup() ([]func() error, error) {
+	var (
+		serviceName = "testing"
+		targetStage = "containerd-snapshotter-base"
+	)
+	pRoot, err := testutil.GetProjectRoot()
+	if err != nil {
+		return nil, err
+	}
+	buildArgs, err := getBuildArgsFromEnv()
+	if err != nil {
+		return nil, err
+	}
+
+	composeYaml, err := testutil.ApplyTextTemplate(composeBuildTemplate, dockerComposeYaml{
+		ServiceName:     serviceName,
+		ImageContextDir: pRoot,
+		TargetStage:     targetStage,
+	})
+	if err != nil {
+		return nil, err
+	}
+	cOpts := []compose.Option{
+		compose.WithBuildArgs(buildArgs...),
+		compose.WithStdio(testutil.TestingLogDest()),
+	}
+
+	return compose.Build(composeYaml, cOpts...)
+
+}
+
+// teardown takes a list of cleanup functions and executes them after integration tests have ended
+func teardown(cleanups []func() error) error {
+	for i := 0; i < len(cleanups); i++ {
+		err := cleanups[i]()
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
