+++ integration/ztoc_test.go
@@ -213,6 +213,7 @@ func TestSociZtocInfo(t *testing.T) {
 			},
 		}
 		for _, tt := range tests {
+			tt := tt
 			t.Run(tt.name, func(t *testing.T) {
 				var zinfo Info
 				output, err := sh.OLog("soci", "ztoc", "info", tt.ztocDigest)
@@ -332,6 +333,7 @@ func TestSociZtocGetFile(t *testing.T) {
 			},
 		}
 		for _, tt := range testCases {
+			tt := tt
 			t.Run(tt.name, func(t *testing.T) {
 				if !tt.expectedErr {
 					for _, f := range files {
